version 1.3.13
update the way to call assertions.js
it can be done by specify, in the shell, the number of the assertion to run
example:

cd assertion && node assertions.js assertion0

added the scripts into the package.json to run the available assertions.

some adjustment to the code of deflate64 are in the making.