v2.1.8-beta ⬇︎ help system redesigned
added flag :
--no-json-out flag

removed flags:
--spawn
--stdout [-p]

begun to work on a dedicated website for a more extensive description, manual, help and examples.
the dedicated website will be developed in react and hosted on Google Cloud Run as a container.

reduced the size of the README.md to basic installation, usage and examples.